Beta Release!! 🎊🥳

+++ src/activity/mod.rs
@@ -25,28 +25,28 @@ pub struct Player {
 pub struct Activity {
     pub name: String,
     pub difficulty: DifficultyOptions,
-    pub rpl: f64,
-    pub cap: f64,
+    pub rpl: u32,
+    pub cap: i32,
     pub player: Player,
 }
 impl Default for Activity {
     fn default() -> Self {
-        let expansion_base = 1350.0;
+        let expansion_base = 1350;
         Activity {
             name: "Default".to_string(),
             difficulty: DifficultyOptions::default(),
             rpl: expansion_base,
-            cap: 100.0,
-            player: Player{pl: expansion_base as u32 +200, class: PlayerClass::default()},
+            cap: 100,
+            player: Player{pl: expansion_base+200, class: PlayerClass::default()},
         }
     }
 }
 impl Activity {
     pub fn get_pl_delta(&self) -> f64 {
-        let gpl = self.player.pl as f64;
+        let gpl = self.player.pl;
         gpl_delta(self.clone(), gpl)
     }
     pub fn get_rpl_mult(&self) -> f64 {
-        rpl_mult(self.rpl)
+        rpl_mult(self.rpl as f64)
     }
 }
